debug-viewer: Use python.install_sources()
With current implementation we would try and install into the system
path regardless of the prefix. On top of that, we could install any left
over pyc file and would install the unit test also.

To fix this, we now list every files to be installed and use
python.install_sources(), leaving to meson the decision on where things
should be installed.