meson: Fix test failures on the CI around nunit-runner
environment().prepend() does not stack. None of the operations on that
object stack. This is documented in the Meson reference manual. Also
avoid overriding previously-set operations because that will cause
a warning in a future version of Meson.

Also, we do not need to manually construct MONO_PATH values. We can
pass an array and have Meson construct it for us.